Added better error state capture to rangeserver-failover-basic tests

+++ tests/integration/rangeserver-failover-basic/run8.sh
@@ -77,6 +77,11 @@ sleep 10
 
 # dump keys
 dump_keys dbdump-a.8
+if [ $? -ne 0 ] ; then
+  kill_all_rs
+  $HT_HOME/bin/stop-servers.sh
+  exit 1
+fi
 
 $HT_HOME/bin/stop-servers.sh
 kill -CONT `cat $HT_HOME/run/Hypertable.RangeServer.rs1.pid`
